Move database form into context form
