+++ src/components/Project/CreateScenarioForms/ContextForm.jsx
@@ -7,9 +7,38 @@ import {
   GENERATE_STREET_CEA,
   GENERATE_TERRAIN_CEA,
 } from './constants';
+import axios from 'axios';
+
+const countryMap = {
+  SG: 'Singapore',
+  CH: 'Switzerland',
+  DE: 'Germany',
+};
+
+const validateDatabase = async (value, databases) => {
+  // Do not need to validate CEA databases
+  if (databases.includes(value)) return Promise.resolve();
+
+  try {
+    const response = await axios.post(
+      `${import.meta.env.VITE_CEA_URL}/api/databases/validate`,
+      {
+        type: 'path',
+        path: value,
+      },
+    );
+    console.log(response);
+    return Promise.resolve();
+  } catch (error) {
+    const errorMessage =
+      error?.response?.data?.detail || 'Unable to validate database.';
+    return Promise.reject(`${errorMessage}`);
+  }
+};
 
 const ContextForm = ({
-  weather,
+  databases = [],
+  weather = [],
   initialValues,
   onChange,
   onBack,
@@ -30,6 +59,43 @@ const ContextForm = ({
       onFinish={onFinish}
       layout="vertical"
     >
+      <Form.Item
+        label="Database"
+        name="database"
+        extra="Select a Database from CEA or link to your own."
+        rules={[
+          { required: true, message: 'Please select a database.' },
+          { validator: (_, value) => validateDatabase(value, databases) },
+        ]}
+      >
+        <Select
+          placeholder="Choose an option from the dropdown"
+          dropdownRender={(menu) => (
+            <div>
+              {menu}
+              <Divider style={{ margin: '4px 0' }} />
+              <OpenDialogButton
+                form={form}
+                name="database"
+                type="directory"
+                placeholder="Or enter path to database folder here"
+              >
+                <FileSearchOutlined />
+                Browse for databases path
+              </OpenDialogButton>
+            </div>
+          )}
+          options={[
+            {
+              label: 'CEA Databases',
+              options: databases.map((database) => ({
+                label: `${database} - ${countryMap?.[database] || database}`,
+                value: database,
+              })),
+            },
+          ]}
+        />
+      </Form.Item>
       <Form.Item
         label="Weather data"
         name="weather"
